+++ package.json
@@ -5,7 +5,8 @@
   "main": "app.js",
   "scripts": {
     "test": "MODE_ENV=test jest --runInBand --detectOpenHandles",
-    "start": "nodemon server.js"
+    "start": "nodemon server.js",
+    "dev" : "nodemon server.js"
   },
   "repository": {
     "type": "git",
