+++ server.js
@@ -1,5 +1,6 @@
 require("dotenv").config();
 const { createApp } = require("./app");
+const { sequelize } = require("./src/models");
 const { redisCli } = require("./src/cache");
 const { settingScore } = require("./src/cache/initScore");
 
@@ -7,7 +8,14 @@ const startServer = async () => {
   const app = createApp();
   const PORT = process.env.PORT;
 
-  // redis-node : 4.5.1 legacyMode
+  await sequelize
+    .sync({ force: false, alter: true })
+    .then(() => {
+      console.log("연결 성공!");
+    })
+    .catch((err) => {
+      console.error(err);
+    });
   await redisCli.connect();
   settingScore();
