+++ src/controllers/userController.js
@@ -2,7 +2,7 @@ const userService = require("../services/userService");
 const error = require("../middlewares/errorConstructor");
 
 const signup = async (req, res) => {
-  const userCreateResult = await userService.create();
+  const userCreateResult = await userService.createUser();
 
   if (!userCreateResult) {
     throw error("ServerError", 500);
@@ -12,7 +12,7 @@ const signup = async (req, res) => {
 
 const getUserInfo = async (req, res) => {
   const { userId } = req.params;
-  const userInfo = await userService.getUserInfo(userId);
+  const userInfo = await userService.getUserInfoById(userId);
   res.status(200).json(userInfo);
 };
