+++ src/routes/userRouter.js
@@ -3,8 +3,8 @@ const router = express.Router();
 const errorHandler = require("../middlewares/errorHandler");
 const userController = require("../controllers/userController");
 
-// router.get("/:userId", errorHandler(userController.getUserInfo));
+router.get("/:userId", errorHandler(userController.getUserInfo));
 
-// router.post("", errorHandler(userController.signup));
+router.post("", errorHandler(userController.signup));
 
 module.exports = router;
